+++ src/Http/Controllers/IssuesController.php
@@ -34,6 +34,7 @@ class IssuesController extends CpController
             'indexUrl' => cp_route('utilities.a11y-report.issues'),
             'criteriaUrl' => cp_route('utilities.a11y-report.criteria'),
             'overviewUrl' => cp_route('utilities.index').'/a11y-report',
+            'settingsUrl' => \Statamic\Facades\Addon::get(\Bpmore\A11yReport\Settings::PACKAGE)?->settingsUrl(),
             'statuses' => IssueQuery::STATUSES,
         ];